requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloup==3.0.5
decorator==5.1.1
ffmpeg-python==0.2.0
Flask==3.0.3
future==1.0.0
glcontext==3.0.0
idna==3.10
isosurfaces==0.1.2
itsdangerous==2.2.0
Jinja2==3.1.4
manim==0.18.1
ManimPango==0.6.0
mapbox_earcut==1.0.2
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
moderngl==5.11.1
moderngl-window==2.4.6
multipledispatch==1.0.0
networkx==3.4.1
numpy==1.26.4
pillow==10.4.0
pycairo==1.27.0
pydub==0.25.1
pyglet==2.0.18
Pygments==2.18.0
pyrr==0.10.3
requests==2.32.3
rich==13.9.2
scipy==1.14.1
screeninfo==0.8.1
skia-pathops==0.8.0.post1
srt==3.5.3
svgelements==1.9.6
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
watchdog==5.0.3
Werkzeug==3.0.4
